Move app router configuration and middleware implementation to the `theme-kit` module
(will be moved to the `theme-kit` package later)
