+++ theme-kit/index.ts
@@ -2,3 +2,4 @@ export { api } from './api';
 export { env, bootHttpEnv } from './env';
 export { createRouter, route } from './router';
 export { locale } from './locale';
+export { middleware } from './middleware';
